Port and remove framework_tests/test_sessions
