+++ osf_tests/test_session.py
@@ -1,6 +1,9 @@
 import pytest
 
-from osf.models import Session
+from framework.sessions import utils
+from tests.base import DbTestCase
+from osf_tests.factories import SessionFactory, UserFactory
+from osf.models import OSFUser as User, Session
 from osf.modm_compat import Q
 
 @pytest.mark.django_db
@@ -27,3 +30,48 @@ class TestSession:
         assert Session.objects.count() == 2  # sanity check
         Session.remove(Q('data.auth_user_id', 'eq', '123ab'))
         assert Session.objects.count() == 1
+
+
+class SessionUtilsTestCase(DbTestCase):
+    def setUp(self, *args, **kwargs):
+        super(SessionUtilsTestCase, self).setUp(*args, **kwargs)
+        self.user = UserFactory()
+        # Ensure usable password
+        self.user.set_password('usablepassword')
+
+    def tearDown(self, *args, **kwargs):
+        super(SessionUtilsTestCase, self).tearDown(*args, **kwargs)
+        User.remove()
+        Session.remove()
+
+    def test_remove_session_for_user(self):
+        SessionFactory(user=self.user)
+
+        # sanity check
+        assert Session.find().count() == 1
+
+        utils.remove_sessions_for_user(self.user)
+        assert Session.find().count() == 0
+
+        SessionFactory()
+        SessionFactory(user=self.user)
+
+        # sanity check
+        assert Session.find().count() == 2
+
+        utils.remove_sessions_for_user(self.user)
+        assert Session.find().count() == 1
+
+    def test_password_change_clears_sessions(self):
+        SessionFactory(user=self.user)
+        SessionFactory(user=self.user)
+        SessionFactory(user=self.user)
+        assert Session.find().count() == 3
+        self.user.set_password('killerqueen')
+        assert Session.find().count() == 0
+
+    def test_remove_session(self):
+        session = SessionFactory(user=self.user)
+        assert Session.find().count() == 1
+        utils.remove_session(session)
+        assert Session.find().count() == 0
